requirements.txt
Django>=4.2.0,<5.0.0
django-environ>=0.11.2
psycopg2-binary>=2.9.9  # PostgreSQL adapter
apache-airflow>=2.7.0,<=2.10.3
django-htmx>=1.17.0
django-tailwind>=3.6.0
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
djangorestframework>=3.14.0
psutil>=5.9.0  # For system metrics collection
